feat: redux list of apis
