+++ src/models/main.ts
@@ -1,6 +1,8 @@
 import {ApiItem, EnvoyFleetItem, StaticRouteItem} from './api';
 
 interface MainState {
+  /** list of apis */
+  apis: ApiItem[];
   /**  currently selected API */
   selectedApi: ApiItem | null;
   /**  currently selected Envoy Fleet */
